requirements.txt
# HARE Skin Lesion Classifier - Requirements
# Final Year Research Project
# Python 3.9+

# Web Framework
streamlit>=1.28.0

# Deep Learning
torch==2.0.1
torchvision==0.15.2
timm==0.9.0

# Image Processing
Pillow>=9.0.0
numpy>=1.21.0

# Visualization
matplotlib>=3.5.0

# ML Utilities
scikit-image>=0.19.0

# Cloud Storage (Optional - for dynamic model loading from Azure)
azure-storage-blob>=12.0.0

# Thesis-demo persistence, authentication, and DICOM support
psycopg[binary]>=3.1.0
passlib[bcrypt]>=1.7.4
bcrypt<4.0.0
pydicom>=2.4.0
